refactor: multiple-select plugin
